refactor(pwa): isolate field primitives for fast refresh
Summary:
- move field primitives into a dedicated ui/fields directory
- split field style exports into a non-TSX helper module
- make the share QR image helper local and exempt main.tsx from the
  fast refresh export rule

Rationale:
- remove react-refresh/only-export-components warnings without changing
  runtime behavior
- keep field-related components organized under a single UI area
- avoid a broad shared ESLint exception by scoping the main.tsx
  override to the PWA package

Tests:
- pnpm --filter @trizum/pwa lint
