+++ packages/pwa/src/ui/fields/DateField.tsx
@@ -8,9 +8,9 @@ import {
   composeRenderProps,
 } from "react-aria-components";
 
-import { cn, type VariantProps } from "./utils";
+import { cn, type VariantProps } from "../utils";
 
-import { fieldGroupVariants } from "./Field";
+import { fieldGroupVariants } from "./Field.styles";
 
 const DateField = AriaDateField;
