+++ packages/pwa/src/ui/fields/Field.tsx
@@ -9,15 +9,8 @@ import {
   type TextProps as AriaTextProps,
   composeRenderProps,
 } from "react-aria-components";
-import { cn, cva, type VariantProps } from "./utils";
-
-const labelVariants = cn([
-  "text-sm font-medium leading-none",
-  /* Disabled */
-  "data-[disabled]:cursor-not-allowed data-[disabled]:opacity-70",
-  /* Invalid */
-  "group-data-[invalid]:text-danger-500",
-]);
+import { type VariantProps, cn } from "../utils";
+import { fieldGroupVariants, labelVariants } from "./Field.styles";
 
 const Label = ({ className, ...props }: AriaLabelProps) => (
   <AriaLabel className={cn(labelVariants, className)} {...props} />
@@ -42,24 +35,6 @@ function FieldError({ className, ...props }: AriaFieldErrorProps) {
   );
 }
 
-export const fieldGroupVariants = cva({
-  variants: {
-    variant: {
-      default: cn([
-        "relative flex h-10 w-full items-center overflow-hidden rounded-md border border-accent-500 dark:border-accent-700 bg-white dark:bg-accent-900 px-3 py-2 text-sm ring-offset-white dark:ring-offset-accent-900",
-        /* Focus Within */
-        "data-[focus-within]:outline-none data-[focus-within]:ring-2 data-[focus-within]:ring-ring data-[focus-within]:ring-offset-2",
-        /* Disabled */
-        "data-[disabled]:opacity-50",
-      ]),
-      ghost: "",
-    },
-  },
-  defaultVariants: {
-    variant: "default",
-  },
-});
-
 interface GroupProps
   extends AriaGroupProps,
     VariantProps<typeof fieldGroupVariants> {}
@@ -75,4 +50,4 @@ function FieldGroup({ className, variant, ...props }: GroupProps) {
   );
 }
 
-export { labelVariants, Label, FieldGroup, FieldError, FormDescription };
+export { Label, FieldGroup, FieldError, FormDescription };
