+++ packages/pwa/src/ui/fields/TextField.tsx
@@ -13,7 +13,7 @@ import {
   Text,
 } from "react-aria-components";
 
-import { cn, type ClassName } from "./utils";
+import { cn, type ClassName } from "../utils";
 
 import { FieldError, Label } from "./Field";
